+++ examples/basic_button.rb
@@ -4,7 +4,7 @@ require 'glimmer-dsl-libui'
 
 include Glimmer
 
-window('hello world', 300, 200, 1) { |w|
+window('hello world', 300, 200) { |w|
   button('Button') {
     on_clicked do
       msg_box(w, 'Information', 'You clicked the button')