+++ examples/basic_window.rb
@@ -4,7 +4,7 @@ require 'glimmer-dsl-libui'
 
 include Glimmer
 
-window('hello world', 300, 200, 1) {
+window('hello world', 300, 200, true) {
   on_closing do
     puts 'Bye Bye'
   end